+++ tests/test_run.py
@@ -183,6 +183,7 @@ def test_run_handler_is_called_on_each_handler():
     registry.register(h)
 
     registry.started(None, None)
+    registry.running(None, None)
     registry.finish(None)
     registry.interrupted(None, None)
     registry.signal_exit()
@@ -202,6 +203,7 @@ def test_run_handler_is_called_on_each_handler():
 
     assert h.calls == [
         "started",
+        "running",
         "finish",
         "interrupted",
         "signal_exit",
@@ -228,6 +230,7 @@ def test_exceptions_does_not_stop_handler_registry():
     registry.register(h)
 
     registry.started(None, None)
+    registry.running(None, None)
     registry.finish(None)
     registry.interrupted(None, None)
     registry.signal_exit()
@@ -247,6 +250,7 @@ def test_exceptions_does_not_stop_handler_registry():
 
     assert h.calls == [
         "started",
+        "running",
         "finish",
         "interrupted",
         "signal_exit",